+++ spec/integrations/pro/consumption/periodic_jobs/long_running_jobs/no_parallel_ticking_when_not_during_pause_spec.rb
@@ -35,8 +35,10 @@ start_karafka_and_wait_until do
 end
 
 # There should be at least one tick parallel to consumption
-assert DT[:consume].all? do |time_range|
+any = DT[:consume].all? do |time_range|
   DT[:ticks].none? do |tick_time|
     time_range.include?(tick_time)
   end
 end
+
+assert any