+++ spec/integrations/rails/rails71_pristine/with-active_job_and_perform_all_later/perform_all_later_run_spec.rb
@@ -11,6 +11,11 @@ require 'tempfile'
 require 'active_job'
 require 'active_job/karafka'
 
+assert_equal(
+  Object,
+  ActiveJob::QueueAdapters::KarafkaAdapter.superclass
+)
+
 ActiveJob::Base.extend ::Karafka::ActiveJob::JobExtensions
 ActiveJob::Base.queue_adapter = :karafka